rag_service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
llama-index-core>=0.14.9
llama-index-embeddings-openai>=0.5.1
PyYAML>=6.0
python-dotenv>=1.0.0
httpx>=0.25.0

